requirements.txt
streamlit==1.31.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
scipy==1.12.0
scikit-learn==1.4.0
matplotlib==3.8.2
plotly==5.18.0  